Fix crons to actually run correctly using the configuration values

+++ config/config.go
@@ -167,10 +167,10 @@ type SystemConfiguration struct {
 	// being sent to the Panel. By default this will send activity collected over the last minute. Keep
 	// in mind that only a fixed number of activity log entries, defined by ActivitySendCount, will be sent
 	// in each run.
-	ActivitySendInterval int64 `default:"60" yaml:"activity_send_interval"`
+	ActivitySendInterval int `default:"60" yaml:"activity_send_interval"`
 
 	// ActivitySendCount is the number of activity events to send per batch.
-	ActivitySendCount int64 `default:"100" yaml:"activity_send_count"`
+	ActivitySendCount int `default:"100" yaml:"activity_send_count"`
 
 	// If set to true, file permissions for a server will be checked when the process is
 	// booted. This can cause boot delays if the server has a large amount of files. In most
